models/requirements.txt
antlr4-python3-runtime==4.8
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.2.0
azure-cognitiveservices-speech==1.27.0
backoff==2.2.1
beautifulsoup4==4.12.2
bitarray==2.7.3
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.34
datasets==2.11.0
decorator==5.1.1
dill==0.3.6
Distance==0.1.3
filelock==3.11.0
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.3.0
g2p-en==2.1.0
httplib2==0.22.0
huggingface-hub==0.13.4
hydra-core==1.0.7
idna==3.4
importlib-metadata==6.2.0
importlib-resources==5.12.0
inflect==6.0.4
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lazy_loader==0.2
librosa==0.10.0.post2
networkx==3.1
nltk==3.8.1
noneprompt==0.1.9
numba==0.56.4
numpy==1.23.5
omegaconf==2.0.6
openai==0.27.4
packaging==23.0
pandas==2.0.0
Pillow==9.5.0
pooch==1.6.0
portalocker==2.7.0
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.7
pyparsing==3.0.9
pypinyin==0.48.0
python-dateutil==2.8.2
python-msspeak==0.6.3
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
responses==0.18.0
rsa==4.9
sacrebleu==2.3.1
scipy==1.10.1
sentencepiece==0.1.97
six==1.16.0
soundfile==0.12.1
soupsieve==2.4
soxr==0.3.5
sympy==1.11.1
tabulate==0.9.0
temp==2020.7.2
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.0
tqdm==4.65.0
transformers==4.27.4
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
wcwidth==0.2.6
xxhash==3.2.0
yarl==1.8.2
zipp==3.15.0aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.8
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.2.0
audioread==3.0.0
azure-cognitiveservices-speech==1.27.0
backoff==2.2.1
beautifulsoup4==4.12.2
bitarray==2.7.3
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.34
datasets==2.11.0
decorator==5.1.1
dill==0.3.6
Distance==0.1.3
dominate==2.7.0
fairseq==0.12.2
filelock==3.11.0
Flask==2.2.3
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.3.0
g2p-en==2.1.0
google==3.0.0
google-api-core==2.11.0
google-api-python-client==2.84.0
google-auth==2.17.2
google-auth-httplib2==0.1.0
google-cloud-language==2.9.1
google-cloud-speech==2.19.0
googleapis-common-protos==1.59.0
grpcio==1.53.0
grpcio-status==1.53.0
httplib2==0.22.0
huggingface-hub==0.13.4
hydra-core==1.0.7
idna==3.4
importlib-metadata==6.2.0
importlib-resources==5.12.0
inflect==6.0.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lazy_loader==0.2
librosa==0.10.0.post2
llvmlite==0.39.1
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.14
networkx==3.1
nltk==3.8.1
noneprompt==0.1.9
numba==0.56.4
numpy==1.23.5
omegaconf==2.0.6
openai==0.27.4
packaging==23.0
pandas==2.0.0
Pillow==9.5.0
pooch==1.6.0
portalocker==2.7.0
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.22.1
psycopg2-binary==2.9.6
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.13
pycparser==2.21
pydantic==1.10.7
pyparsing==3.0.9
pypinyin==0.48.0
python-dateutil==2.8.2
python-msspeak==0.6.3
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
responses==0.18.0
rsa==4.9
sacrebleu==2.3.1
scikit-learn==1.2.2
scipy==1.10.1
sentencepiece==0.1.97
six==1.16.0
soundfile==0.12.1
soupsieve==2.4
soxr==0.3.5
SQLAlchemy==2.0.9
sqlalchemy-cockroachdb==2.0.0
sympy==1.11.1
tabulate==0.9.0
temp==2020.7.2
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.0
torchaudio==2.0.1
tqdm==4.65.0
transformers==4.27.4
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
visitor==0.1.3
voices==1.2.0
wcwidth==0.2.6
Werkzeug==2.2.3
xxhash==3.2.0
yarl==1.8.2
zipp==3.15.0
